Begin fixler-spelling.json
I'm just starting to feel out the waters here. All of this stuff is
messy, and Wikipedia is a bit, too. In the large table I found here, and
which I'm using for all of this work (to start, anyway):

https://en.wikipedia.org/wiki/List_of_Latin-script_alphabets

...the section titled:

Letter–diacritic combinations: connected or overlaid

...contains the Latin letter A with ogonek, and links to its page, which
has a disambiguation "not to be confused with" link to the Cyrillic
version, which is not out in the main table on the link above. This
makes me wonder how many more things aren't out in the main table, and
how I'll imagine to discover them.

Either way, this means I also now have the issue of two different
letters A-with-ogonek, so I need some way to allow both in a way that
isn't ambiguous. I'm also running into three different "hooks": the
palatalized hook, retroflex hook, and rhotic hook, each of which has its
own page, and the main page has a long table of characters that use the
various hooks. However, that page links further to a hook above page,
for Vietnamese characters. There's a lot to figure out just in hooks.

I'm starting out with 5 bits of info: the name, which isn't always
clear, even from the title of the Wikipedia page, so I'm using a mix of
things, including the actual Unicode name of symbol, or diacritic, to
help there; the majuscule and minuscule forms, which are currently pairs
of the base Latin letter, and the accented form; the diacritics, which
come in a list currently, as, e.g., the Vietnamese letters can have
diacritics that stack; and a link, which I'll be using to link each
thing back to the best Wikipedia article I can.

These 5 bits are in little, unnamed dictionaries, which are in order in
a list. This list, to start with, and named for the section in the
Wikipedia article linked at the top here, is for the connected diacritic
combinations. I'm currently creating a variable for each diacritic, but
I think I'll move to a dictionary, so the names of each are also easily
discoverable.