Add more letters to connected list
I stopped here to add the note that I've run into a conflict, I'm sure
the first of many. The African D has two issues: 1) it features a
different diacritic in its majuscule (stroke) and minuscule (hook)
forms, and 2) in its majuscule form, it conflicts with the existing
entry for D with stroke.

One way around the first issue is to simply allow different majuscule
and minuscule diacritics lists. I'd like to keep the simplicity of a
single list of diacritics (a list because of things like Vietnamese
stackable diacritics), but allow for, e.g., a missing list, replaced by
one for majuscule and minuscule, say as minusculeDiacritics, etc.

There's an issue I want to keep first and foremost, which this has
underscored, which is that almost no one is an expert on all of these
diacritics, and the reason for even wanting to use them is typically
going to be because you want to mimic a word you've seen, or because
someone in a steno reel somewhere has asked "can you write X using
steno?" You don't want to have to become intimately familiar with the
letter+accent before being able to write it, say, through copying it
with your mouse and Googling for it. You want to try out what looks like
it'll work on your steno board, and watch it magically Just Work. I
don't know that I can 100% do that, but it's a goal. I realize this
makes, e.g., all the various hooks an issue.

For now, I've left the African D a TODO entry.