Add diagonal stroke versus bar concept
I don't love this, but I did say I wanted to make this work more
visually than anything else. Now I encounter the L with stroke, and L
with bar entries, and so far, "stroke" has meant a horizontal stroke.
This stroke is diagonal, and the bar is horizontal. So, I can either
keep up the name of the thing, where stroke just means diagonal now, and
then have to come up with another thing that means bar (where the
horizontally-stroked, barred I in an earlier entry, I just treated as
using a stroke), or I can stick with the visual side, use the stroke
chord for the barred L here, and then come up with another thing for the
visually diagonal stroke in the L with stroke. I've opted, for now, for
the latter, and added in a stroke_diagonal diacritic to try to keep up
both the visual side, and the name side of things.

Maybe, eventually, in an effort to be all things to all people, I'll
have a flag that lets you set whether you're a stickler for details, and
want to use all the correct terminology, so this diagonally-stroked L
will use the stroke chord, or if you want to go visual, and let the
shape, position, and angle decide what chord gets used, as I'm currently
defaulting to.