Add more entries, change diagonal stroke to slash
I've decided, for now (I reserve the right to change my mind yet again),
that the diagonal stroke should just be called—at least in code—a slash.
I think, in the eventual documentation, it'll be described in some way
as something like a visually diagonally-upward slash, even though the
name may be stroke, or bar, or whatever.