Merge diacritics and modifications into modifiers
To simplify stacking accents and modifiers, like schwa with a cedilla,
which needs to invert the e, then add the diacritic, I've decided that
everything is just modifications to the original Latin letter. I didn't
want to have to juggle looking for several things in different places,
and I think this still makes sense for the library.