Move uncertain entries into TBD dictionary
This was the modifieds dictionary, but I did away with that distinction,
and now I think this is just a home for all the things I'm working on,
that don't belong in the main list[s] yet, for one reason or another,
like that I think Cyrillic entries will need their own system, as will
things like voiced fricatives. I'll understand better how to serve these
letters when I'm further along in this plugin.