Resurrect æ entries, and add to letters dictionary
This adds the idea that a letter can be something more than just A-Z, so
it can be looked up and assembled easily. I still have to figure out how
the minuscule vs majuscule stuff will work, though. I also added an
example of a "turned" modifier, which has no majuscule, so I put in
None. I could also just not have that key, and check for the existence
of both minuscule and majuscule, but I sort of like the intentionality
of this method, so it doesn't look like an oversight.