Add a very messy first pass of generating output
This is just me working through some of how I'll do this. There are many
things still to consider, like minuscule vs majuscule, capitalization
stuff (like {^}, and {-|}), and glue. I'm also not sure why there needs
to be connected and detached dictionaries, except that I used Wikipedia
to get a lot of entries, and it split them up. I'm going to be using the
Unicode Latin blocks to fill in a bunch of things the page I was using
seemed to skip over. I'm still getting my head around these characters.