Rename letters dict, add all capital letters
After much thought, I realize the best bet for getting capitals to work,
and leaving things open for new ideas later, is to simply add all the
capital letters to the letters dict, along with the previously added
lowercase letters. This will let me put in minuscule and majuscule
things that aren't in the alphabet, as witnessed by the inclusion
already of some Æ/æ entries. I also realized that "letters" is the wrong
name, and switched it to characters, which should open up thoughts about
other things for use here later.