Build out entire dictionary of entries finally
This creates both the minuscule and majuscule entries. I had to list the
modStrokes in the createOutlines function, because the map object it was
creating was dying somehow after the minuscule line, but before the
majuscule line, so all majuscules were mapping to empty string.

I'm using a pair of wrap strings for wrapping the majuscule and
minuscule in Plover things to ensure upper and lowercase. The output
isn't valid json data yet, though, especially because of the single
quote strings python useses.