Update cedilla and turned strokes
I kept forgetting the hook shape I had made for the cedilla, and
confusing it with the hook and horn (which both still also confuse me),
and I just wanted it to be the single character down low that it used to
be in my non-plugin JSON file, so I'm just making it the −G for now.

I also decided I like −RL instead of −FG for the turned modifier stroke.
It's the same thing—opposite corners of the 6 diacritic keys—but it's a
bit easier for my hand to stroke −RL than −FG. It's a me thing.