Add several, and fixup existing modifiers
I decided that ring should be ringAbove, because it's always spoken of
that way in Unicode. I also added a bunch of below modifiers, some
doubled modifiers, and an inverted modifier, in preparation for adding a
bunch of new symbols from another large list I found on Wikipedia:

https://en.wikipedia.org/wiki/List_of_precomposed_Latin_characters_in_Unicode

Before, I had the double acute, and the double grave simply created as
repeated entries in the modifiers lists in the associated entries, but I
realized that this meant the idea of these features was not centralized.
If I wanted to change them later, I'd have to find and replace them
throughout the file. By naming them in the modifiers dictionary, and
simply doubling the strokes in the string values there, it makes it easy
for myself and others to quickly change how they're created, just like
the new breveInverted, and all the new below entries. I decided that the
belows for many of these will also just double the original stroke,
because their symbols are too big to be moved in the six keys I'm using
for diacritics, unlike dotAbove and dotBelow. Also, they're rare, so I
don't mind a rare doubling of strokes.

I don't love doubling the macron for macronBelow, but the obvious stroke
for macron below, -RB, is already the stroke diacritic. I'll think about
it for a bit. Make I'll move the stroke.