Fix up some incorrect entries
I found around 4 where I just put the wrong accented character in. There
could be more, but we'll find them eventually. I also realized that the
E with ring above didn't really exist, so I deleted it. I was looking it
up to see if the upper and lowercase versions existed, and I ended up
hitting backspace in the search box, and removed the ring on the E with
ring above that I had pasted in there from this file. Turns out the one
I got online was using a combining mark, and there isn't a precomposed
version in Unicode. I also added a TODO comment for a second version of
a lowercase s with diacritic that I ran into.