Merge all entry dictionaries together
It feels like they should all just be in the same place. The Vietnamese
one was only some of the characters from that alphabet, which weren't
already in the others. I never cared about the connected versus detached
dichotomy. If anything, it might be interesting to tag them as part of
the various languages and such, to allow fine grain search, and other
uses, but I'll leave that to the future, and probably someone else.