Change how generating a stroke image works
I've decided to use the EU keys for certain things, rather than doubling
up strokes, e.g. for "below", or "doubled" versions of diacritics. As
such, there's no longer a clear up/down pair of states, because I'll
need to do new colors for the new section, so I'm getting rid of the
"boolean blindness" of the "pressed" arg, and replacing it with the
nasty "keyType" string arg.