Add remaining links, remove J with caron majuscule
Turns out there is no J with caron majuscule. It was another combined
character. I wrote a unicode thing the other day to search all my
characters for decomposed letters, and it found none, so I guess I did
it wrong, because I've run into several now.

Also, a lot of letters, especially this whole set I added at the end,
from another set on Wikipedia, doesn't seem to have pages on Wikipedia,
so I've currently opted to link them to compart.com, which looks like a
major business, which makes me not want to link to them, but the data
seems simple, good, and all the characters are there, uniformly.