Change schwa chord back to what it used to be
I wasn't thinking. I had it as SKHA for the word, SKHA* for the symbol,
and I added SKHA*P for the upper case version I didn't know existed
before this project. Then I realized that SKWHA was slightly easier for
me to reach, and felt more like the spelling. I don't know if I realized
this when I was initially adding it, but SKWHA is using the Emily's
Symbols unique starter, and the left uncuddle. Adding the star just
capitalizes next, and adding the P outputs a grave. So, back it goes.