Move tweaks into their own column in diacritics
I want to try having the tweak part of the diacritics images not as part
of the chord image, but as it's own separate chord in a second column.
It'll make the rows more even, and make it easier to search for them,
and maybe talk about them in the column a bit, calling them out better.