Add all characters list to README.md
I'm still playing with this idea, but I finally, after a couple of days,
figured out a decent collation of these characters, which uses the base
characters, lowercased, followed by the list of diacritics, if any,
followed by if the base character is lowercase, as a boolean. There are
so many ways this could be sorted, but I just wanted something that
looked unoffensive, and even a bit pleasing, and this is that for me.

I admit, I used ChatGPT to help me a lot in finding info about unicode
collation, and had it write me the initial sorting function, based on my
criteria, but it was wrong, and didn't work, so I had to pick it apart,
and learn how it worked, which I would always do, anyway, but it saved
me some time spiraling on it.