Move reversed C and turned F back into TBD
I also moved the AE ligature turned entry into a new section that I
haven't committed yet, so it appears to have been deleted in this
commit.

I had a little break through last night about how I want to handle
things like turned C with acute (or whatever it was I saw), and the AE
ligature turned, or with various diacritics.

I'm currently printing out a JSON dictionary that I build up at the end
of everything in the program. I've only been handling one [large] set of
diacritics, however, and some turns and ligatures. I have alternate
dictionaries I want, however, like Greek, and Morse. Those need to be
added in, too, and counted in the docs (where it says how many
characters the dictionary currently supports). I also need to output the
base characters, like A through Z, which I'm not currently doing. I've
been pondering how I want to add everything into the dictionary...

I've also run into the issue of multiple changes to a base character,
like a turned letter that also has a diacritic, and of multiple
characters, like AE → Æ. Further, I've run into both of these things,
like the ᴂ (AE ligature turned), and Ǽ (AE ligature with acute). To get
around the Æ examples, I just added Æ to the dictionary of modifiable
characters, a clear hack, which doesn't work for everything. I've been
looking for a way to handful all of this neatly.

I've decided that I just need to start with the base dictionary of
English letters, and order the entries such that as they get created,
they add to the dictionary, for use by later entries. I also need to
have entries work with multiple letters. This gives me a path where I
can remove the AE ligature entries from the modifiable letters
dictionary, and finally have an AE ligature entry that has, e.g.,
"minuscule": ("ae", "æ"), with "modifiers": ["ligature"], the output of
which then gets added back into the growing dictionary of what will be
exported as the final fingerspelling dictionary. Then, the AE ligature
turned entry can use, e.g., "minuscule": ("æ", "ᴂ"), with "modifiers":
["turned"].

This idea, of a base dictionary, with the starting letters in it, which
then gets added to by other processes, creating parts which can then be
used by later entries, filling up a dictionary that can be looked
through by the docs creator, seems to solve all my issues. It's not the
only thing; because outputs get wrapped in upper and lowercase Plover
meta things, and because the dictionary keys are actually the strokes,
it means I need a parallel dictionary to hold mappings of characters to
strokes, which can then be looked up, and I might want to add the
characters to their own dictionaries, too, for individual sorting and
docs purposes, but I think I've solved my main problems.