Introduce idea of core dictionary
This is a step toward creating a core dictionary, into which each
character, as it's created, gets added, so later entries can use them.
For example, a ligature can be created in one entry, with the composed
character added back to the core dictionary, and then a version of that
ligature with a diacritic can start from that ligature, and just add the
diacritic to it, like it's one of the starting, Latin letters.