Remove minuscule I with dot above
Turns out there is no minuscule I with dot above. Unicode just points
this to U+0069, the small i. This messes with my little system, where
each created character gets put back into a core dictionary, mapped to
its strokes, so the next thing to come along needing that character will
find the created version. The problem is that this overwrites the
original i, telling it to use the outline *EU/-F, for I with dot above,
minuscule. This means all the modified i's that come later, and only the
minuscule ones, now start with that outline, so when I tried to use ī
tonight, and couldn't get it to work, I would have had to actually
stroke *EU/-F/-FP, instead of *EU/-FP for minuscule I with macron. I
can't even just move it to the end, because it would still mess up the
plain old minuscule i. Computers... I give them up.