Move AE ligatures to new system, remove dupes
I think I stated in an earlier commit that I liked my old idea of
putting characters back into a growing dictionary as I created the
strokes for them, so you could have an AE → Æ ligature entry, which
would put the Æ back into the pot, and then a later Æ → Ǽ entry could
pull that Æ out of the pot, and turn it into the Ǽ. This idea seemed
cool at the time, but had a few glaring issues.

For one, I ran into an issue with small i with dot above, which doesn't
actually exist, and which showed up in some online Unicode lists as just
the letter i. I don't know if Unicode has some kind of assignment here,
but it tripped up my system, because it would run into i with dot above,
which uses the -F chord for the dot above, which would then get put back
into the system as just the i, which would overwrite i in the
dictionary, so the outline would change from *EU to *EU/-F, and then all
other things that needed the i, like ī, î, and ï, would pull that out,
and now those three would be *EU/-F/-FP, *EU/-F/-RPG, and *EU/-F/-FL.

I also didn't like mutating the original dictionary, even though I ended
up working on a copy of it, which also felt like a hack. I didn't like
that I had to write so much explanation in comments, like this one, in
my code, to explain why I was doing all of this. It also started to
intrude on some ideas I had with the Russian alphabet to allow various
outlines to represent certain letters, which I could see as either
phonetic, or orthographic in nature, and wanted to have both, like the
Cyrillic letter Х, which I could see (with -RPG as my unique Russian
alphabet ender), as KP-RPG, because it looks just like an X, or as
KHARPG, after ха, its Russian name, which transliterates to Kha, its
Wikipedia page name. It also got in the way of having alternate ways of
writing the English alphabet, like STKPW* versus STK* for "z".

Furthermore, it works out better for the docs to have all modifiers
listed in each entry. If I have an AE → Æ entry, and it's separate from
the Æ → Ǽ entry, then the "Used in:" lists after each modifier don't
pick up everything. The acute entry lists Ǽ, but the ligature entry
doesn't, because Ǽ is created from the base Æ, and doesn't have in its
modifiers list that it actually needs the ligature in it's train of
strokes. I wasn't looking forward to building out a system to spider
through the train of these, and I think, because of the other reasons,
it wouldn't have been worth it anyway.