Annotate some weird entries
Now that I got rid of the core dictionary system, I can add back in the
weird, but sort of correct small I with dot above entry, because it was
weird that it didn't work, and output an untrans. It's still weird, but
in a different way that I don't mind as much.