Clean up docs for modifiers, change cedilla stroke
I decided I didn't like the stroke for the cedilla. I always stumbled
trying to write it, and I missed when I had it (before formalizing
everything in this system) as the -B, and, in this system for a time, as
the -G. So, I'm moving it back to -B, but adding in the "extra" tweak. I
kind of want to make the cedilla just the -B, and move comma below to
EB, but I just really like comma, the stroke from Emily's Symbols, being
the comma. I probably shouldn't, but I currently do.