Simplify counting characters in docs
I've been wanting to do this for a while now. Instead of manually
counting things in loops in the docs generator, I wanted to just get the
entire, created dictionary, and ask for its length. This should do a
much better job of keeping the count up to date with changes now.