Add alphabet-building function
This is based on an as-yet-uncommitted, new version of the Latin letters
alphabet, which has more data for each letter, and also groups majuscule
and minuscule together in a single entry, which is a little dictionary.
I'm also now building a LUT of the dictionary, which maps just the
letter to it's basic, composed stroke (i.e., without case directives and
such), for an easy way to be able to reference any character, which I'll
need now, because now the Latin letters dictionary is a list thereof, so
I won't be able to directly look up letters now, for building all the
Latin letters with diacritics, which is the point of this system.