Change the way the Latin alphabet system works
The whole point of this change is to add more info to each letter,
mostly for documentation, and to factor out the minuscule and majuscule
chords, so they can be defined in a variable each, outside all of the
alphabet data, to make it easy to change them all in one go. This latter
bit is also why I built the whole stroke parse/merge/render set of
functions, to make it easy to add unique enders to letter bases. If I
were to build a steno system, I'd make it much easier to play with
chords in an expression-based way, without all this need for parsing
things, and checking for issues. It's all pretty sloppy.