Update alphabet format
This changes the Latin, Greek, and Russian alphabets, renaming them from
alphabet to alphabetData, to explain better what they're for, moves
their majuscule and minuscule strokes inside, to contain and simplify
passing them into functions with their alphabets, renames "outline" to
"strokes" for the alphabetData entries that map to the majuscule and
minuscule parts of the letter strokes, and, finally, puts all letter
strokes—which are now just the letter part ("A", not "A*" or "A*P")—into
lists, which I thought I didn't want at this level, but which answers an
issue I had for a while, where I didn't know what to do about letters
that different people stroke differently, like STKPW and STK for Z/z.

I like having the unique, factored-out minuscule/majuscule parts in the
dictionaries now, to group them, but I also sort of don't like it,
because it bundles them up, making it a little bit harder to set
separately. Before, they were just globals, so you could set them in the
module, which I normally don't want, but felt easy. At some point, I
want users to be able to override these in their own user file. Before,
that would just be `latinMinEnder = etc.`, whereas now it'll need to be
something like `latinAlphabetData["minStroke"] = etc.`, which feels like
less of a setting, than it does just coding. I may go with the former
name, and somehow have it pull from the user setting, if it exists, at
alphabet-building time. We'll see.