Rewrite buildModdedChar → buildModCharOutlines
I changed how everything in here works. It now takes an alphabet, and
doesn't take wraps. It returns a tuple of the destination character—
because it simplifies things at the call site to do so—along with a list
of all the outlines that create that character, each as a list of
strokes, which later need to be joined. I may change that last bit, if
it turns out easier to join them with slashes before returning here.