Remove chi alternate stroke in the Greek alphabet
Xi, is the unaspirated "ks" sound, like the "x" in English, but doesn't
look like an X. Chi looks like an X, but doesn't sound like one. Because
I was using the [steno fingerspelling] X (KP) for the Chi, I couldn't
use it for an orthographic alternate for Xi, and was using a ch (KH)
sound. However, later on, I forgot this, and happily added XP in as an
alternate for Xi. Then I couldn't write, it when I was trying to add
some more Greek things to this system. Now it's fixed.

I also removed from a Russian character's docs its cool connection with
Chi (because it, too, looked like an X, and uses it as an
orthographic alternate), as it no longer shares any such connection.