Fix tests against recent changes
I should have had these fixes in with the original changes, but I'm not
become to fighting form yet. I decided a bit late to add in tests, so
I'm not doing this my usually TDD way, which means I have a ton of
little issues all over the place, and little confidence in how it all
works right now, particularly because it's actually currently broken.