Build list of dictionary bits as constants
Not real constants; the Python kind, where you just name them with
capital letters, and hope. The final dictionary builder has been broken
for many commits now, as I've been giving this thing a major overhaul.
This isn't how I would have things in the past; it's a lot more messy.
I'm looking forward to a return to form. These bits are needed in the
working, but as of yet uncommitted, final dictionary builder.