Rewrite all of buildFingerSpellingDict
As I've changed basically every other part of this file, this bit, which
uses all of those, needed a huge overhaul itself. It's been broken for
many commits now, completely, which is not how I normally work, but
whatever. We're just going seat of the pants for a bit, I guess.