Version all new modifier images, and a changed one
Cedilla changed quite a while ago (days? a couple of weeks?), but I've
only just noticed I never put it in. I also added 7 additional
modifiers, because they were needed for a bunch of knew Greek things I
entered the past couple of days. The docs and dictionary are not
up-to-date yet, but they will be soon.