Add stroke how-tos to readme all-characters list
I struggled to figure out where to put this for a while. I kept coming
back to the idea. There really isn't a good place, but I think this
works. Now, you can look through the "Used in:" lists, after each
modifier in the modifiers section, which features all the modified
characters, which all link to the all-characters list at the bottom, and
then when you jump down, you can see which strokes are needed, or at
least, from one layer of indirection back. It doesn't tell you that you
need A*P/-RPG/-FPBG for Ẫ, but rather A + circumflex + tilde. This way,
if there are several strokes for A in your alphabet, you can just use
the one you like. It also doesn't spoon-feed you the strokes; you have
to actually learn the strokes for circumflex and tilde to use these
how-tos. I don't feel bad, though, because the strokes are memorable.

The one issue I forsee is that this doesn't allow for different ways of
getting to a character. I'm trying to avoid such things, like I'm
playing with currency symbol entries at the moment, and there was one
that has a symbol, but informally uses a letter with diacritics in upper
or lowercase. I had uppercase open in the entry, and started to put the
regular letter in there, figuring why not, but then realized I didn't
like the inconsistency of it. You can still write both versions of that
letter with diacritics, and none of the other currency symbols allowed a
similar hack, so I got rid of it.

Another thing I need to get to is the regular alphabet entries. They're
not yet call out in the readme. So when one of these how-tos says to
stroke a Greek character, followed by a diacritic (not sure any do yet,
but they will, and so will Russian), you don't yet know, via the readme,
how to do that.