Add code points to alphabet letters, with links
I still want a way to share the Wikipedia links I have hiding in the
link property of countless characters. I'd also like to remove the ones
that just link, e.g., somewhat ineffectually, to the modifier wiki page,
like some that just link to to the cedilla, or tilde. I also need to add
links to tons that don't have them yet.