Massively rewrite all of imgsgen.py
I did not develop this in my usually way, through TDD, with careful,
granular updates. This was a total, seat-of-the-pants free-for-all, so
I'm committing what I have now, so I can move forward more safely. It's
able to draw the full keyboard now. I'm about to make it so it can draw
a subset of the keys, for use in more easily breaking out sections.