Fix Dotless I/J, i/j with dot above mess
I think this is satisfactory. There aren't dotted versions of i and j,
because they already have dots, and there aren't dotless versions of I
and J, because they don't have dots to begin with, so I'm free to use
the dot above modifier here to add dots (formally, "tittles") to the
majuscule  letters, and to remove them from the minuscule letters.