Allow for caseless character mappings
I've been using minuscule and majuscule for a while now, without issue,
other than the weirdness of not always having both. However, now I've
run into the issue of mixed-case ligatures, like ǈ, where if I list it
as a minuscule, then it gets forced by the Plover case formatting to the
ligature ǉ, and if majuscule, then it gets forced to Ǉ. I don't love the
solution, and it does slow things down, if not noticeably, because
there's a ton of checking now for the existence of the "caseless" key in
every modified character dictionary, but I can live with it... for now.