Add in combining diacritics system
Originally, I had this working from a separate dictionary, but I've
liked the idea of combining the combining diacritics entries in my
system with the actual combining diacritics characters themselves.

I'm not entirely sure what I'll do if I end up with combiners that I
don't have in the system, but there are several options there.